+++ tech1-framework-domain/src/main/java/io/tech1/framework/domain/tuples/TupleMonetaryUnit.java
@@ -1,0 +1,6 @@
+package io.tech1.framework.domain.tuples;
+
+import java.math.BigDecimal;
+
+public record TupleMonetaryUnit(BigDecimal amount, BigDecimal price) {
+}