+++ RELEASE_NOTES.md
@@ -7,6 +7,7 @@
 - Add utilities: SpringAuthoritiesUtility, TimestampUtility (methods: getStartOfMonthTimestamp)
 - Add postgres converters: Version
 - Add postgres persistable abstractions
+- Add tuples: TupleMonetaryUnit
 
 ### v1.16
 - Docker "servers" (tech1-framework-b2b-mongodb-server, tech1-framework-b2b-postgres-server) → GitHub Packages